requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0
httpx>=0.25.0
python-docx>=1.0.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
oracledb>=2.0.0
